backend/requirements.txt
absl-py==2.1.0
alembic==1.13.3
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.0
cycler==0.12.1
decorator==4.4.2
dlib==19.24.6
et-xmlfile==1.1.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.0
filelock==3.16.1
flatbuffers==24.3.25
flower==2.0.1
fonttools==4.54.1
fsspec==2024.10.0
h11==0.14.0
httptools==0.6.1
humanize==4.11.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
jax==0.4.35
jaxlib==0.4.35
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
kombu==5.4.2
lz4==4.3.3
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.2
mediapipe==0.10.18
ml_dtypes==0.5.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
proglog==0.1.10
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.1.0
psycopg2==2.9.9
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
redis==5.1.0
requests==2.32.3
scenedetect==0.6.4
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
SQLAlchemy==2.0.35
starlette==0.38.6
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.39
ultralytics-thop==2.0.12
urllib3==2.2.3
uvicorn==0.31.0
uvloop==0.20.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
